Make checks for existence of ThreadLocal more robust (#1136)
We should try to load an actual value of a ThreadLocal, because
in some environments (like WebLogic) a `NoClassDefException` won't
be thrown during loading of `JdkProvider`.

We should also catch all possible errors, not only `NoClassDefException`
to avoid a situation when the provider becomes a null reference.

(cherry picked from commit cced96a)