Fix a racy read bug in EWMA
Because `rate` is only volatile, but not an atomic primitive, its
updates are racy. The problem with the current code is that we read
`rate` twice before making a  write. That means we could calculate a
change relying on one rate, but add it to another. The fix is to read
`rate` to a local variable and calculate a new value based on it. This
guarantees that the new rate is based only one rate and produces a
correct value. We still have a "lost update" problem, but it's not
important, because it doesn't matter how many threads update the rate
simultaneously. They all set the same new rate based on the current
count.