Improve metrics filter (#1193)
* Don't use a ClockProvider

`Clock` is already an abstraction over a tick represented by a long value,
so it seem superfluous to use a provided for a class which doesn't have
any state and doesn't change.

* Make TimerRequestEventListener a static inner class

So we its creation doesn't carry all fields of
`InstrumentedResourceMethodApplicationListener` and we explicitly
set fields its state.

* Remove the superfluous method "key"

* Split contexts to several variables

To be more much more clear what event handlers do. Otherwise, they
reuse the common variable and the correct behaviour depends on the
sequence in which Jersey produces events.

* Add explicit flag to control tracking filters

We should not guess whether to include filter metrics based on the
metric name. Since tracking of filter times is a new feature, it
should be enabled explicitly for users who want it. For other users,
the process of migrating to the new instrumented listener should be
backward-compatible.