Introduce Summing to keep track of sum of events (#1254)
This is a rework of PR #1022 by @hashbrowncipher (and fixes @slovdahl's
issue #712). This PR is based on the 4.0-development branch, which has
evolved further than master.

The main differences are:

* Introduction of a new Summing interface, which is implemented by
Meter, Timer and Histogram
* Reporting of the Summing.sum() is done as a long to avoid loss of precision on
high values due to floating point properties
* All reporters report the sum() property of the relevant metrics

The main purpose of storing the sum is to facilitate monitoring systems
that use raw ascending counters (e.g. Prometheus), and time-series
databases such as InfluxDB (see also @mevdschee comment)