Add Metrics 3/4 legacy adapter (#1258)
This adds an adapter module which should allow software which is
depended Metrics 3/4 to coexist together with Metrics 5.0.

The idea to allow users to run their Dropwizard/Spring Boot apps with
Metrics 5.0. The adapter will provide the API compatible with the
old Metrics API which delegates requests to the new module.

Old integrations and reporters will work, and at the same time it
should possible to implement new reporters to take advantages
of tags for metric names and the sum of measurements for Timers.