ExponentiallyDecayingReservoirTest improvements (#1587)
Improves some of the comments and test parameters in `com.codahale.metrics.ExponentiallyDecayingReservoirTest`.

I think that https://github.com/dropwizard/metrics/issues/1582 can be closed after this is merged as the implementation is correct and statistical error is expected.

The increase of alpha value for `longPeriodsOfInactivityShouldNotCorruptSamplingState` should eliminate/reduce failure. I couldn't recrate this value.